requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
ollama>=0.4.0
google-cloud-secret-manager>=2.0.0
google-cloud-storage>=2.0.0
msal>=1.28.0
requests>=2.28.0
